+++ Chips/Chips.jsx
@@ -64,7 +64,11 @@ class ChipIcon extends Icon {
       otherprops['tabindex'] = 0;
       otherprops['role'] = 'button';
     }
-    return <Icon {...props} {...otherprops}>{children}</Icon>;
+    return (
+      <Icon {...props} {...otherprops}>
+        {children}
+      </Icon>
+    );
   }
 }