+++ webpack.config.js
@@ -37,8 +37,5 @@ module.exports = {
   },
   plugins: [
     new CssMigrationWebpackPlugin(),
-    new WebpackShellPlugin({
-      onBuildEnd: ['node compileComponents.js']
-    })
   ]
 };