+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ## 1.4.3
+
 * Update to mdc 0.35.0.
+* Addded Typography component.
 * Changed stroked prop to outlined in Button.
 * Changed stroked prop to outlined in Card.
 * Added entry chips.